requirements.txt
ccxt
pandas
numpy
matplotlib
google-genai
requests
